Refactor test runner
